Refactor deployment of common contracts
We expose the DeployContracts method to the public, and allow users
to pass in their contracts of choice, for deployment.
We also expose the CommonContracts global variable, which
pre-configures the most common contracts.
We also introduce a new option, called WithContracts, in order to
programmatically specify which contracts should be deployed by the
Blockchain.